collect all matplotlib-realted functions to sim_data_plot.py
